InventoryReportCommandHandler
